feat: Refactor UI components and constants
Extracts constants to `lib/constants.ts` and introduces `Header` and `Sidebar` components to improve UI organization. Updates `App.tsx` to utilize these new components and `lib/render2d.ts` for 2D rendering logic.
